reduced the amount of samples used for testing
