--- csemlib/data/topography/col.txt
@@ -1,5 +1,0 @@
-9.000000000000000000e+01
-9.000000000000000000e+01
-9.000000000000000000e+01
-9.000000000000000000e+01
-9.000000000000000000e+01
