--- csemlib/data/topography/topo_short.txt
@@ -1,5 +1,0 @@
--180.000000, 90.000000, -4228
--179.833333, 90.000000, -4228
--179.666667, 90.000000, -4228
--179.500000, 90.000000, -4228
--179.333333, 90.000000, -4228
